Fix to migration.

+++ db/migrate/20110505143820_remove_tax_booker_setting.rb
@@ -1,7 +1,7 @@
 class RemoveTaxBookerSetting < ActiveRecord::Migration
   def self.up
     ci = ConfigurationItem.where(:key => 'tax_booker_class_name').first
-    ci.destroy
+    ci.destroy if ci
   end
 
   def self.down
